+++ CHANGELOG.md
@@ -64,6 +64,8 @@
   and the second returns the number of entries in the warehouse.
 - Changed the `add_column` method. It raises a `ValueError` if the x-offset of the column is not unique. 
 - Changed `src` folder name in `automatic_warehouse`.
+- Changed `resources` folder name in `automatic_warehouse-res`.
+- Changed `configuration` folder name in `automatic_warehouse-config`.
 - `Drawer` name changed to `Tray`.
 - `Deposit` name changed to `Bay`.